fix(error-tracking): revert modal logics to explicit declarations for kea-typegen
kea-typegen uses static analysis and can't see actions/reducers/selectors
injected by dynamic factory functions. Revert the three modal logics to
explicit declarations while keeping the shared component deduplication
(RuleModal, RuleList, SortableRuleItem).
